requirements/requirements-api.txt
pip==1.5.6
djangorestframework
django
mysqlclient
django-rest-swagger
stevedore
colorlog
six
kombu